+++ src/App.tsx
@@ -7,6 +7,7 @@ import TestBB from './pages/TestBB';
 import TestSS from './pages/TestSS';
 import TestVS from './pages/TestVS';
 import TestSM from './pages/TestSM';
+import TestMB from './pages/TestMB';
 import Header from './components/Header/Header';
 
 function App() {
@@ -20,6 +21,7 @@ function App() {
                 <Route path="/testvs" element={<TestVS />} />
                 <Route path="/testbb" element={<TestBB />} />
                 <Route path="/testss" element={<TestSS />} />
+                <Route path="/testmb" element={<TestMB />} />
             </Routes>
         </BrowserRouter>
     );
