+++ src/pages/TestMB.tsx
@@ -1,0 +1,7 @@
+import Login from '../components/Login';
+
+function TestMB() {
+    return <Login />;
+}
+
+export default TestMB;
